fix some build issues
